requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.0
gunicorn==21.2.0